+++ imod/mf6/drn.py
@@ -1,7 +1,6 @@
-import itertools
 from dataclasses import asdict
 from datetime import datetime
-from typing import Dict, Optional
+from typing import Optional
 
 import numpy as np
 
@@ -34,20 +33,7 @@ from imod.schemata import (
 )
 from imod.typing import GridDataArray
 from imod.typing.grid import enforce_dim_order, is_planar_grid
-
-
-def expand_repetitions(
-    repeat_stress: list[datetime], time_min: datetime, time_max: datetime
-) -> Dict[datetime, datetime]:
-    expanded = {}
-    for year, date in itertools.product(
-        range(time_min.year, time_max.year + 1),
-        repeat_stress,
-    ):
-        newdate = date.replace(year=year)
-        if newdate < time_max:
-            expanded[newdate] = date
-    return expanded
+from imod.util.expand_repetitions import expand_repetitions
 
 
 class Drainage(BoundaryCondition, IRegridPackage):
@@ -182,7 +168,7 @@ class Drainage(BoundaryCondition, IRegridPackage):
         cls,
         key: str,
         imod5_data: dict[str, dict[str, GridDataArray]],
-        period_data: dict[str, dict[str, GridDataArray]],
+        period_data: dict[str, list[datetime]],
         target_discretization: StructuredDiscretization,
         target_npf: NodePropertyFlow,
         allocation_option: ALLOCATION_OPTION,
