+++ imod/tests/test_mf6/test_mf6_dis.py
@@ -197,7 +197,7 @@ def test_write_ascii_griddata_2d_3d(idomain_and_bottom, tmp_path):
 @pytest.mark.usefixtures("imod5_dataset")
 def test_from_imod5_data__idomain_values(imod5_dataset):
     imod5_data = imod5_dataset[0]
-    period_data = imod5_dataset[1]
+
     dis = imod.mf6.StructuredDiscretization.from_imod5_data(imod5_data)
 
     # Test if idomain has appropriate count
@@ -209,7 +209,7 @@ def test_from_imod5_data__idomain_values(imod5_dataset):
 @pytest.mark.usefixtures("imod5_dataset")
 def test_from_imod5_data__grid_extent(imod5_dataset):
     imod5_data = imod5_dataset[0]
-    period_data = imod5_dataset[1]
+
     dis = imod.mf6.StructuredDiscretization.from_imod5_data(imod5_data)
 
     # Test if regridded to smallest grid resolution
